makes text gray, creates new designable view file
